// IMSProj/src/app/Interfaces/purchaseDt.ts
export interface PurchaseDt{

      createdBy: string
      createdDate: string
      createdWorkstation: string
      modifiedBy: string
      modifiedDate: string
      modifiedWorkstation: string
      id: number
      purchaseid: number
      storeCode: string
      itemCode: string
      qty: number
      unitPrice: number
      totalPrice: number

}